refactor: use abstract class `PathfinderMob` and interface `InventoryCarrier` in data provider
