+++ src/main/kotlin/InventoryAccessor.kt
@@ -1,0 +1,9 @@
+package dev.kpherox.vihp
+
+import net.minecraft.world.entity.npc.InventoryCarrier
+import net.minecraft.world.item.Items
+
+object InventoryAccessor {
+  fun getInventory(entity: InventoryCarrier) =
+      entity.getInventory().getItems().filterNot { it.`is`(Items.AIR) }
+}
